+++ packages/components/chip/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/adevinta/spark/compare/@spark-ui/chip@1.2.4...@spark-ui/chip@1.2.5) (2023-11-22)
+
+**Note:** Version bump only for package @spark-ui/chip
+
 ## [1.2.4](https://github.com/adevinta/spark/compare/@spark-ui/chip@1.2.3...@spark-ui/chip@1.2.4) (2023-10-30)
 
 **Note:** Version bump only for package @spark-ui/chip